+++ ui/src/components/editor/Tools/ClearTool/connect.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux';
 
 import { ClearTool } from './ClearTool';
-import { selectEditorClearMode } from '../../../../../store/selectors';
-import { setClearMode } from '../../../../../store/editorActions';
+import { selectEditorClearMode } from '../../../../store/selectors';
+import { setClearMode } from '../../../../implementations/kapture-backend/actions/editorActions';
 
 const mapState = (state, props) => {
   return { clearMode: selectEditorClearMode(state) };
