+++ ui/src/components/index.js
@@ -8,3 +8,4 @@ export { PlateTabBar } from './editor/PlateTabBar';
 export { Plate } from './editor/Plate';
 export { NoPlatesMessage } from './editor/Plate';
 export { ComponentList } from './editor/ComponentList';
+export { Tools } from './editor/Tools';
